+++ plugins/guests/redhat/cap/os_variant.rb
@@ -3,11 +3,11 @@ module VagrantPlugins
     module Cap
       class OsVariant
         def self.os_variant(machine)
-          command = "grep VARIANT_ID /etc/os-release"
+          command = "grep VARIANT_ID #{OS_RELEASE_FILE}"
           # TODO: execute efficient command to solve this
           if machine.communicate.test(command) # test if command is exits with code 0
-            machine.communicate.execute(command) do |type, data|
-              return data.chomp.gsub(/"/, '').split("=").last
+            machine.communicate.execute(command) do |_, data|
+              return data.chomp.delete('"').split('=').last
             end
           end
         end